requirements.txt
streamlit==1.32.0
groq==0.9.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
numpy==1.26.3
gtts==2.3.2
elevenlabs==0.2.26
edge-tts==6.1.9
pyyaml==6.0.1
num2words==0.5.14
streamlit-extras==0.3.5
plotly==5.18.0
langchain==0.1.0
langchain-groq==0.0.5
firebase-admin==6.3.0
streamlit-autorefresh==1.0.1
streamlit-player==0.4.7
streamlit-authenticator==0.2.3
streamlit-lottie==0.0.5
streamlit-timeline==0.0.2
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-card==0.0.7
nltk==3.8.1
scikit-learn==1.3.2
icecream==2.1.3
psutil==5.9.6
folium==0.15.0
streamlit-folium==0.15.0
qrcode==7.4.2
emoji==2.8.0